Guarded preload table

+++ inc/Engine/Preload/Subscriber.php
@@ -9,6 +9,7 @@ use WP_Rocket\Engine\Preload\Controller\ClearCache;
 use WP_Rocket\Engine\Preload\Controller\LoadInitialSitemap;
 use WP_Rocket\Engine\Preload\Controller\Queue;
 use WP_Rocket\Engine\Preload\Database\Queries\Cache;
+use WP_Rocket\Engine\Preload\Database\Tables\Cache as CacheTable;
 use WP_Rocket\Event_Management\Subscriber_Interface;
 use WP_Rocket_Mobile_Detect;
 
@@ -64,17 +65,25 @@ class Subscriber implements Subscriber_Interface {
 	protected $queue;
 
 	/**
+	 * Cache table.
+	 *
+	 * @var CacheTable
+	 */
+	protected $table;
+
+	/**
 	 * Creates an instance of the class.
 	 *
-	 * @param Options_Data            $options Options instance.
-	 * @param LoadInitialSitemap      $controller controller creating the initial task.
-	 * @param Cache                   $query Cache query instance.
-	 * @param Activation              $activation Activation manager.
+	 * @param Options_Data $options Options instance.
+	 * @param LoadInitialSitemap $controller controller creating the initial task.
+	 * @param Cache $query Cache query instance.
+	 * @param Activation $activation Activation manager.
 	 * @param WP_Rocket_Mobile_Detect $mobile_detect Mobile detector instance.
-	 * @param ClearCache              $clear_cache Clear cache controller.
-	 * @param Queue                   $queue Preload queue.
+	 * @param ClearCache $clear_cache Clear cache controller.
+	 * @param Queue $queue Preload queue.
+	 * @param CacheTable $table Cache table.
 	 */
-	public function __construct( Options_Data $options, LoadInitialSitemap $controller, $query, Activation $activation, WP_Rocket_Mobile_Detect $mobile_detect, ClearCache $clear_cache, Queue $queue ) {
+	public function __construct( Options_Data $options, LoadInitialSitemap $controller, $query, Activation $activation, WP_Rocket_Mobile_Detect $mobile_detect, ClearCache $clear_cache, Queue $queue, CacheTable $table ) {
 		$this->options       = $options;
 		$this->controller    = $controller;
 		$this->query         = $query;
@@ -82,6 +91,7 @@ class Subscriber implements Subscriber_Interface {
 		$this->mobile_detect = $mobile_detect;
 		$this->clear_cache   = $clear_cache;
 		$this->queue         = $queue;
+		$this->table = $table;
 	}
 
 	/**
@@ -121,6 +131,11 @@ class Subscriber implements Subscriber_Interface {
 	 * @return void
 	 */
 	public function maybe_load_initial_sitemap( $old_value, $value ) {
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		if ( ! isset( $value['manual_preload'], $old_value['manual_preload'] ) ) {
 			return;
 		}
@@ -144,6 +159,11 @@ class Subscriber implements Subscriber_Interface {
 	 * @return void
 	 */
 	public function maybe_cancel_preload( $old_value, $value ) {
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		if ( ! isset( $value['manual_preload'], $old_value['manual_preload'] ) ) {
 			return;
 		}
@@ -166,6 +186,11 @@ class Subscriber implements Subscriber_Interface {
 	 */
 	public function update_cache_row() {
 		global $wp;
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		$url = home_url( add_query_arg( [], $wp->request ) );
 
 		if ( $this->query->is_preloaded( $url ) ) {
@@ -200,6 +225,11 @@ class Subscriber implements Subscriber_Interface {
 	 */
 	public function delete_url_on_not_found() {
 		global $wp;
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		$url = home_url( $wp->request );
 		$this->query->delete_by_url( $url );
 	}
@@ -210,6 +240,11 @@ class Subscriber implements Subscriber_Interface {
 	 * @return void
 	 */
 	public function on_permalink_changed() {
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		$this->query->remove_all();
 		$this->controller->load_initial_sitemap();
 	}
@@ -246,6 +281,10 @@ class Subscriber implements Subscriber_Interface {
 	 */
 	public function clean_url( string $url ) {
 
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		$this->clear_cache->partial_clean( [ $url ] );
 	}
 
@@ -255,6 +294,11 @@ class Subscriber implements Subscriber_Interface {
 	 * @return void
 	 */
 	public function clean_full_cache() {
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		set_transient( 'wpr_preload_running', true );
 		$this->queue->add_job_preload_job_check_finished_async();
 		$this->clear_cache->full_clean();
@@ -269,6 +313,11 @@ class Subscriber implements Subscriber_Interface {
 	 * @return void
 	 */
 	public function clean_partial_cache( $object, array $urls, $lang ) {
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		// Add Homepage URL to $purge_urls for preload.
 		$urls[] = get_rocket_i18n_home_url( $lang );
 
@@ -284,6 +333,10 @@ class Subscriber implements Subscriber_Interface {
 	 */
 	public function clean_urls( array $urls ) {
 
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		$this->clear_cache->partial_clean( $urls );
 	}
 
@@ -294,6 +347,11 @@ class Subscriber implements Subscriber_Interface {
 	 * @return void
 	 */
 	public function delete_post_preload_cache( $post_id ) {
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		if ( ! $this->options->get( 'manual_preload', 0 ) ) {
 			return;
 		}
@@ -315,6 +373,10 @@ class Subscriber implements Subscriber_Interface {
 	 */
 	public function delete_term_preload_cache( $term_id ) {
 		if ( ! $this->options->get( 'manual_preload', 0 ) ) {
+			return;
+		}
+
+		if(! $this->table->exists()) {
 			return;
 		}
 
@@ -342,6 +404,11 @@ class Subscriber implements Subscriber_Interface {
 	 * }
 	 */
 	public function preload_after_automatic_cache_purge( $deleted ) {
+
+		if(! $this->table->exists()) {
+			return;
+		}
+
 		if ( ! $deleted || ! $this->options->get( 'manual_preload' ) ) {
 			return;
 		}
